Fix table when clicking 'Show visible to students'
Previously, when the 'Show visible to students' option was
checked, a reduced list of todo items would be shown in the table,
but not the right item. The items would be selected from
the front of the un-reduced list. Now, the right todo items
should show when the 'Show visible to students' option is
checked.